% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getCovariatesTable,TrenaProject-method}
\alias{getCovariatesTable,TrenaProject-method}
\alias{getCovariatesTable}
\title{Get the covariates table in which each sample, for which expression data is available, is described}
\usage{
\S4method{getCovariatesTable}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\value{
A data.frame
}
\description{
Get the covariates table in which each sample, for which expression data is available, is described
}
