% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getEncodeDHS,TrenaProjectHG38-method}
\alias{getEncodeDHS,TrenaProjectHG38-method}
\alias{getEncodeDHS}
\title{Get all the dnase hypersensitivity regions in the expansive region covered by the enhancer}
\usage{
\S4method{getEncodeDHS}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
Get all the dnase hypersensitivity regions in the expansive region covered by the enhancer
}
