% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getEnhancerTissues,TrenaProjectHG38-method}
\alias{getEnhancerTissues,TrenaProjectHG38-method}
\alias{getEnhancerTissues}
\title{Get all the enhancer tissues included in the current genehancer}
\usage{
\S4method{getEnhancerTissues}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
Get all the enhancer tissues included in the current genehancer
}
\seealso{
getEnhancers
}
