% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getExpressionMatrix,TrenaProject-method}
\alias{getExpressionMatrix,TrenaProject-method}
\alias{getExpressionMatrix}
\title{Get the a specifc expression matrix}
\usage{
\S4method{getExpressionMatrix}{TrenaProject}(obj, matrixName)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{matrixName}{A numeric matrix}
}
\description{
Get the a specifc expression matrix
}
