% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getExpressionMatrixNames,TrenaProjectHG38-method}
\alias{getExpressionMatrixNames,TrenaProjectHG38-method}
\alias{getExpressionMatrixNames}
\title{Get the names of the expression matrices - their names with directory and .RData suffix stripped out}
\usage{
\S4method{getExpressionMatrixNames}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
Get the names of the expression matrices - their names with directory and .RData suffix stripped out
}
