% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getFootprintDatabaseHost,TrenaProjectHG38-method}
\alias{getFootprintDatabaseHost,TrenaProjectHG38-method}
\alias{getFootprintDatabaseHost}
\title{get the host on which the footprints Postgres database is running}
\usage{
\S4method{getFootprintDatabaseHost}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get the host on which the footprints Postgres database is running
}
