% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getFootprintDatabaseNames,TrenaProjectHG38-method}
\alias{getFootprintDatabaseNames,TrenaProjectHG38-method}
\alias{getFootprintDatabaseNames}
\title{get the names of the database tables holding footprints}
\usage{
\S4method{getFootprintDatabaseNames}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get the names of the database tables holding footprints
}
