% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getFootprintDatabasePort,TrenaProjectHG38-method}
\alias{getFootprintDatabasePort,TrenaProjectHG38-method}
\alias{getFootprintDatabasePort}
\title{get the port  on which the footprints Postgres database is running}
\usage{
\S4method{getFootprintDatabasePort}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get the port  on which the footprints Postgres database is running
}
