% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getGeneInfoTable,TrenaProjectHG38-method}
\alias{getGeneInfoTable,TrenaProjectHG38-method}
\alias{getGeneInfoTable}
\title{return the data.frame with gene ids, chromosome, tss of the primary transcript, and strand for all genes in the project}
\usage{
\S4method{getGeneInfoTable}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\value{
a data.frame with gene names as rownames
}
\description{
return the data.frame with gene ids, chromosome, tss of the primary transcript, and strand for all genes in the project
}
