% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getGeneRegulatoryRegions,TrenaProjectHG38-method}
\alias{getGeneRegulatoryRegions,TrenaProjectHG38-method}
\alias{getGeneRegulatoryRegions}
\title{Get gene regulatory regions for the current target gene, or one explicitly named, using the specified strategy}
\usage{
\S4method{getGeneRegulatoryRegions}{TrenaProjectHG38}(obj,
  targetGene = NA, tissues = "all", fallback.upstream = 5000,
  fallback.downstream = 5000)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}

\item{targetGene}{default NA, in which case the current object's targetGene is used.}
}
\description{
Get gene regulatory regions for the current target gene, or one explicitly named, using the specified strategy
}
\seealso{
setTargetGene
}
