% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getGenome,TrenaProjectHG38-method}
\alias{getGenome,TrenaProjectHG38-method}
\alias{getGenome}
\title{get the standard short name (e.g., "hg38", "mm10") of the project under study}
\usage{
\S4method{getGenome}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get the standard short name (e.g., "hg38", "mm10") of the project under study
}
