% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getPrimaryTranscriptInfo,TrenaProjectHG38-method}
\alias{getPrimaryTranscriptInfo,TrenaProjectHG38-method}
\alias{getPrimarytTranscriptInfo}
\title{Get ENSEMBL-derived genomic information on what they judget to be the primary transcript of this gene}
\usage{
\S4method{getPrimaryTranscriptInfo}{TrenaProjectHG38}(obj,
  targetGene = NA)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}

\item{targetGene}{default NA, in which case the previously assigned targetGene is assumed}
}
\description{
Get ENSEMBL-derived genomic information on what they judget to be the primary transcript of this gene
}
\seealso{
setTargetGene
}
