% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getSupportedGenes,TrenaProjectHG38-method}
\alias{getSupportedGenes,TrenaProjectHG38-method}
\alias{getSupportedGenes}
\title{get the list of genes supported in this project}
\usage{
\S4method{getSupportedGenes}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get the list of genes supported in this project
}
