% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getTargetGene,TrenaProject-method}
\alias{getTargetGene,TrenaProject-method}
\alias{getTargetGene}
\title{get the single gene currently designated for analysis}
\usage{
\S4method{getTargetGene}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the single gene currently designated for analysis
}
