% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getTranscriptsTable,TrenaProject-method}
\alias{getTranscriptsTable,TrenaProject-method}
\alias{getTranscriptsTable}
\title{Get the transcripts for the gene}
\usage{
\S4method{getTranscriptsTable}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
Get the transcripts for the gene
}
