% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getVariantDataset,TrenaProjectHG38-method}
\alias{getVariantDataset,TrenaProjectHG38-method}
\alias{getVariantDataset}
\title{Get the specified variants table}
\usage{
\S4method{getVariantDataset}{TrenaProjectHG38}(obj, datasetName)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}

\item{datasetName}{character string, the variant dataset of interest}
}
\description{
Get the specified variants table
}
