% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getVariantDatasetNames,TrenaProjectHG38-method}
\alias{getVariantDatasetNames,TrenaProjectHG38-method}
\alias{getVariantDatasetNames}
\title{List the RData files in the variants directory}
\usage{
\S4method{getVariantDatasetNames}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
List the RData files in the variants directory
}
