% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneHancerDB.R
\docType{methods}
\name{listTissues,GeneHancerDB-method}
\alias{listTissues,GeneHancerDB-method}
\alias{listTissues}
\title{Return a character vector containing all of the tissues known to GeneHancer}
\usage{
\S4method{listTissues}{GeneHancerDB}(obj)
}
\arguments{
\item{obj}{An object of class GeneHancerDB}
}
\description{
Return a character vector containing all of the tissues known to GeneHancer
}
