% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{recognizedGene,TrenaProject-method}
\alias{recognizedGene,TrenaProject-method}
\alias{recognizedGene}
\title{Do we have expression data for the suggested gene? genomic and epigenetic information?}
\usage{
\S4method{recognizedGene}{TrenaProject}(obj, geneName)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{geneName}{A character string in the same protocol as the project's expression matrices}
}
\value{
a chrom.loc (chrom:start-end) string
}
\description{
Do we have expression data for the suggested gene? genomic and epigenetic information?
}
