% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{setTargetGene,TrenaProjectHG38-method}
\alias{setTargetGene,TrenaProjectHG38-method}
\alias{setTargetGene}
\title{Set a single gene for analysis}
\usage{
\S4method{setTargetGene}{TrenaProjectHG38}(obj, targetGene,
  curatedGenesOnly = FALSE)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}

\item{targetGene}{a characteor string, the name of the gene}
}
\description{
Set a single gene for analysis
}
