% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{show,TrenaProjectHG38-method}
\alias{show,TrenaProjectHG38-method}
\alias{show}
\title{get a summary of the objects}
\usage{
\S4method{show}{TrenaProjectHG38}(object)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get a summary of the objects
}
