% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{class}
\name{TrenaProject-class}
\alias{TrenaProject-class}
\alias{.TrenaProject}
\alias{TrenaProject}
\title{TrenaProject}
\usage{
TrenaProject(supportedGenes, genomeName, footprintDatabaseHost,
  footprintDatabaseNames, expressionDirectory, variantsDirectory,
  covariatesFile, quiet)
}
\arguments{
\item{genomeName}{A string indicating the genome used by the Trena object.
Currently, only human and mouse ("hg38","mm10") are supported}

\item{quiet}{A logical indicating whether or not the Trena object should print output}

\item{geneSymbol}{A character string in standard HUGO nomenclature}

\item{expressionDataDirectory}{A string pointing to a collection of RData expression matrices}
}
\value{
An object of the TrenaProject class
}
\description{
TrenaProject and its (projected) subclasses provide convenient containers in which to collect
 trena-related aggregation of a gene's (a hierarchy of classes) including expression data,
transcript and variant info, genomic and epigenomic context, trena models and/or the means to create them
}
\examples{
# Create a Trena object using the human hg38 genome
cola1a <- TrenaProject("COL1A1", "hg38")

}
