% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\name{getChipSeq,TrenaProject-method}
\alias{getChipSeq,TrenaProject-method}
\alias{getChipSeq}
\title{Get all the ReMap 2018 ChIP-seq binding sites in the extended (i.e., enhancers-extended) region of the gene}
\usage{
\S4method{getChipSeq}{TrenaProject}(obj, chrom, start, end, tfs = NA)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{chrom}{string vector}

\item{start}{numeric vector}

\item{end}{numeric vector}

\item{tfs}{one of more tfs - limit the hits to those for this transcription factor/s}
}
\value{
data.frame
}
\description{
Get all the ReMap 2018 ChIP-seq binding sites in the extended (i.e., enhancers-extended) region of the gene
}
