% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getCovariatesTable,TrenaProjectHG38-method}
\alias{getCovariatesTable,TrenaProjectHG38-method}
\alias{getCovariatesTable}
\title{Get the covariates table in which each sample, for which expression data is available, is described}
\usage{
\S4method{getCovariatesTable}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\value{
A data.frame
}
\description{
Get the covariates table in which each sample, for which expression data is available, is described
}
