% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\name{getEncodeDHS,TrenaProject-method}
\alias{getEncodeDHS,TrenaProject-method}
\alias{getEncodeDHS}
\title{Get all the dnase hypersensitivity regions in the expansive region covered by the enhancer}
\usage{
\S4method{getEncodeDHS}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
Get all the dnase hypersensitivity regions in the expansive region covered by the enhancer
}
