% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getEnhancers,TrenaProjectHG38-method}
\alias{getEnhancers,TrenaProjectHG38-method}
\alias{getEnhancers}
\title{Get all the enhancer regions for the gene}
\usage{
\S4method{getEnhancers}{TrenaProjectHG38}(obj,
  targetGene = NA_character_, tissues = "all")
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}

\item{targetGene}{default NA, in which case the current object's targetGene is used.}
}
\description{
Get all the enhancer regions for the gene
}
\seealso{
setTargetGene
}
