% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getEnhancers,TrenaProject-method}
\alias{getEnhancers,TrenaProject-method}
\alias{getEnhancers}
\title{Get all the enhancer regions for the gene}
\usage{
\S4method{getEnhancers}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
Get all the enhancer regions for the gene
}
