% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getExpressionMatrix,TrenaProjectHG38-method}
\alias{getExpressionMatrix,TrenaProjectHG38-method}
\alias{getExpressionMatrix}
\title{Get the a specifc expression matrix}
\usage{
\S4method{getExpressionMatrix}{TrenaProjectHG38}(obj, matrixName)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}

\item{matrixName}{A numeric matrix}
}
\description{
Get the a specifc expression matrix
}
