% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getExpressionMatrixNames,TrenaProject-method}
\alias{getExpressionMatrixNames,TrenaProject-method}
\alias{getExpressionMatrixNames}
\title{Get the names of the expression matrices - their names with directory and .RData suffix stripped out}
\usage{
\S4method{getExpressionMatrixNames}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
Get the names of the expression matrices - their names with directory and .RData suffix stripped out
}
