% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getFootprintDatabaseHost,TrenaProject-method}
\alias{getFootprintDatabaseHost,TrenaProject-method}
\alias{getFootprintDatabaseHost}
\title{get the host on which the footprints Postgres database is running}
\usage{
\S4method{getFootprintDatabaseHost}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the host on which the footprints Postgres database is running
}
