% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getFootprintDatabaseNames,TrenaProject-method}
\alias{getFootprintDatabaseNames,TrenaProject-method}
\alias{getFootprintDatabaseNames}
\title{get the names of the database tables holding footprints}
\usage{
\S4method{getFootprintDatabaseNames}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the names of the database tables holding footprints
}
