% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getGeneEnhancersRegion,TrenaProject-method}
\alias{getGeneEnhancersRegion,TrenaProject-method}
\alias{getGeneEnhancersRegion}
\title{Get the chromosomal region enclosing the enhancers of the current targetGene, with a flanking percentage added}
\usage{
\S4method{getGeneEnhancersRegion}{TrenaProject}(obj, flankingPercent = 0)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{flankingPercent}{a numeric percentage of the gene's total span}
}
\value{
a chrom.loc (chrom:start-end) string
}
\description{
Get the chromosomal region enclosing the enhancers of the current targetGene, with a flanking percentage added
}
