% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getGeneInfoTable,TrenaProject-method}
\alias{getGeneInfoTable,TrenaProject-method}
\alias{getGeneInfoTable}
\title{return the data.frame with gene ids, chromosome, tss of the primary transcript, and strand for all genes in the project}
\usage{
\S4method{getGeneInfoTable}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\value{
a data.frame with gene names as rownames
}
\description{
return the data.frame with gene ids, chromosome, tss of the primary transcript, and strand for all genes in the project
}
