% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getGeneRegion,TrenaProject-method}
\alias{getGeneRegion,TrenaProject-method}
\alias{getGeneRegion}
\title{Get the chromosomal region surrounding the current targetGene, with a flanking percentage added up and downstream}
\usage{
\S4method{getGeneRegion}{TrenaProject}(obj, flankingPercent = 0)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{flankingPercent}{a numeric percentage of the gene's total span}
}
\value{
a chrom.loc (chrom:start-end) string
}
\description{
Get the chromosomal region surrounding the current targetGene, with a flanking percentage added up and downstream
}
