% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getGenome,TrenaProject-method}
\alias{getGenome,TrenaProject-method}
\alias{getGenome}
\title{get the standard short name (e.g., "hg38", "mm10") of the project under study}
\usage{
\S4method{getGenome}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the standard short name (e.g., "hg38", "mm10") of the project under study
}
