% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getPrimaryTranscriptInfo,TrenaProject-method}
\alias{getPrimaryTranscriptInfo,TrenaProject-method}
\alias{getPrimarytTranscriptInfo}
\title{Get ENSEMBL-derived genomic information on what they judget to be the primary transcript of this gene}
\usage{
\S4method{getPrimaryTranscriptInfo}{TrenaProject}(obj, targetGene = NA)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{targetGene}{default NA, in which case the previously assigned targetGene is assumed}
}
\description{
Get ENSEMBL-derived genomic information on what they judget to be the primary transcript of this gene
}
\seealso{
setTargetGene
}
