% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getSupportedGenes,TrenaProject-method}
\alias{getSupportedGenes,TrenaProject-method}
\alias{getSupportedGenes}
\title{get the list of genes supported in this project}
\usage{
\S4method{getSupportedGenes}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
get the list of genes supported in this project
}
