% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getTargetGene,TrenaProjectHG38-method}
\alias{getTargetGene,TrenaProjectHG38-method}
\alias{getTargetGene}
\title{get the single gene currently designated for analysis}
\usage{
\S4method{getTargetGene}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
get the single gene currently designated for analysis
}
