% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProjectHG38.R
\docType{methods}
\name{getTranscriptsTable,TrenaProjectHG38-method}
\alias{getTranscriptsTable,TrenaProjectHG38-method}
\alias{getTranscriptsTable}
\title{Get the transcripts for the gene}
\usage{
\S4method{getTranscriptsTable}{TrenaProjectHG38}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProjectHG38}
}
\description{
Get the transcripts for the gene
}
