% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getVariantDataset,TrenaProject-method}
\alias{getVariantDataset,TrenaProject-method}
\alias{getVariantdataset}
\title{Get the specified variants table}
\usage{
\S4method{getVariantDataset}{TrenaProject}(obj, datasetName)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{datasetName}{character string, the variant dataset of interest}
}
\description{
Get the specified variants table
}
