% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{getVariantDatasetNames,TrenaProject-method}
\alias{getVariantDatasetNames,TrenaProject-method}
\alias{getVariantDatasetNames}
\title{List the RData files in the variants directory}
\usage{
\S4method{getVariantDatasetNames}{TrenaProject}(obj)
}
\arguments{
\item{obj}{An object of class TrenaProject}
}
\description{
List the RData files in the variants directory
}
