% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TrenaProject.R
\docType{methods}
\name{setTargetGene,TrenaProject-method}
\alias{setTargetGene,TrenaProject-method}
\alias{setTargetGene}
\title{Set a single gene for analysis}
\usage{
\S4method{setTargetGene}{TrenaProject}(obj, targetGene,
  curatedGenesOnly = FALSE)
}
\arguments{
\item{obj}{An object of class TrenaProject}

\item{targetGene}{a character string, the name of the gene}
}
\description{
Set a single gene for analysis
}
